requirements.txt
# Core dependencies for HF Spaces
gradio>=5.9.1
python-dotenv>=1.0.0
numpy>=1.24.3
typing-extensions>=4.0.0

# Audio processing and analysis
pydub>=0.25.1
soundfile>=0.12.1
librosa>=0.10.1
ffmpeg-python>=0.2.0

# OpenAI integration
openai>=1.58.1

# Hugging Face deployment
huggingface-hub>=0.20.0

# Additional dependencies for audio processing
scipy>=1.9.0
scikit-learn>=1.1.0
